--- fdr_hacking/__init__.py
@@ -1,5 +1,0 @@
-"""Top-level package for FDR Hacking."""
-
-__author__ = """Chakravarthi Kanduri"""
-__email__ = 'chakra.kanduri@gmail.com'
-__version__ = '0.0.1'
