--- fdr_hacking/cli.py
@@ -1,13 +1,0 @@
-"""Console script for fdr_hacking."""
-# todo
-
-
-import typer
-
-
-def main():
-    return
-
-
-if __name__ == "__main__":
-    typer.run(main)
