+++ scripts/cli/calculate_hist_results.py
@@ -1,6 +1,5 @@
-import argparse
 import ast
-from fdr_hacking.data_generation import *
+from scripts.analysis.data_generation import *
 import numpy as np
 import pandas as pd
