+++ scripts/cli/plot_histograms.py
