+++ scripts/cli/statistical_test.py
@@ -1,9 +1,9 @@
 import argparse
 from itertools import product
-from fdr_hacking.statistical_testing import *
+from scripts.analysis.statistical_testing import *
 import numpy as np
 import pandas as pd
-from fdr_hacking.util import parse_yaml_file
+from scripts.analysis.util import parse_yaml_file
 
 
 def execute():
