--- scripts/some_example.py
@@ -1,6 +1,0 @@
-
-
-def test():
-    # this test will run when you run make test-all
-    # only test in files ending with _example.py in this directory will run
-    assert True
