+++ setup.py
@@ -31,13 +31,13 @@ setup(
     description=" ",
     entry_points={
         'console_scripts': [
-            'simulate_data=scripts.cli_scripts.simulate_data:execute',
-            'statistical_test=scripts.cli_scripts.statistical_test:execute',
-            'calculate_hist_results=scripts.cli_scripts.calculate_hist_results:execute',
-            'plot_histograms=scripts.cli_scripts.plot_histograms:execute',
-            'plot_comparative_boxcharts=scripts.cli_scripts.plot_comparative_boxcharts:execute',
-            'simulate_semi_real_world_data=scripts.cli_scripts.simulate_semi_real_world_data:execute',
-            'execute_statistical_test=scripts.cli_scripts.execute_statistical_test:execute'
+            'simulate_data=scripts.cli.simulate_data:execute',
+            'statistical_test=scripts.cli.statistical_test:execute',
+            'calculate_hist_results=scripts.cli.calculate_hist_results:execute',
+            'plot_histograms=scripts.cli.plot_histograms:execute',
+            'plot_comparative_boxcharts=scripts.cli.plot_comparative_boxcharts:execute',
+            'simulate_semi_real_world_data=scripts.cli.simulate_semi_real_world_data:execute',
+            'execute_statistical_test=scripts.cli.execute_statistical_test:execute'
         ],
     },
     install_requires=requirements,
@@ -46,7 +46,7 @@ setup(
     include_package_data=True,
     keywords='fdr_hacking',
     name='fdr_hacking',
-    packages=find_packages(include=['fdr_hacking', 'fdr_hacking.*', 'cli_scripts', 'scripts.*', 'pipelines']),
+    packages=find_packages(include=['scripts.*']),
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/kanduric/fdr_hacking',
